+++ nosql_schema/post_processors.py
@@ -1,0 +1,7 @@
+"""
+This module contains the post processor definitions
+"""
+
+
+def bool_processor(value):
+    return bool(value)